Reach 100% line and branch test coverage
Add tests for Avo resources, filters, actions and controllers, application
controllers and concerns, helpers, mailers, models, queries, services, the
rubyandrails.info import classes and rake tasks.

Remove code no test could reach: ActionCable channel boilerplate,
ApplicationJob, User#can_edit?, and unreachable blank guards in
AuthorProposal and the search query objects. Move rake task bodies into
FtsTables and AdminUserCreator per docs/llms/rake.md.

Bugs surfaced by the new tests and fixed: course price conversion used
Float (19.99 became 1998 cents); BlockBannedRequests ran before
Authentication set Current.user, so suspended users were never blocked.

+++ app/controllers/entries_controller.rb
@@ -242,7 +242,7 @@ class EntriesController < ApplicationController
 
     # Convert price from dollars to cents
     if all_permitted_params[:price].present?
-      params_hash[:price_cents] = (all_permitted_params[:price].to_f * 100).to_i
+      params_hash[:price_cents] = (all_permitted_params[:price].to_d * 100).to_i
     end
 
     params_hash
